+++ gazebo/physics/Link.hh
@@ -171,11 +171,15 @@ namespace gazebo
 
       /// \brief Set the linear acceleration of the body.
       /// \param[in] _accel Linear acceleration.
-      public: void SetLinearAccel(const ignition::math::Vector3d &_accel);
+      /// \deprecated acceleration should be achieved by setting force, see SetForce()
+      public: void SetLinearAccel(const ignition::math::Vector3d &_accel)
+              GAZEBO_DEPRECATED(9.0);
 
       /// \brief Set the angular acceleration of the body.
       /// \param[in] _accel Angular acceleration.
-      public: void SetAngularAccel(const ignition::math::Vector3d &_accel);
+      /// \deprecated acceleration should be achieved by setting force, see SetForce()
+      public: void SetAngularAccel(const ignition::math::Vector3d &_accel)
+              GAZEBO_DEPRECATED(9.0);
 
       /// \brief Set the force applied to the body.
       /// \param[in] _force Force value.
@@ -679,9 +683,11 @@ namespace gazebo
       protected: Visuals_M visuals;
 
       /// \brief Linear acceleration.
+      /// deprecated
       protected: ignition::math::Vector3d linearAccel;
 
       /// \brief Angular acceleration.
+      /// deprecated
       protected: ignition::math::Vector3d angularAccel;
 
       /// \brief Offsets for the attached models.