+++ Changelog.md
@@ -12,6 +12,9 @@
 1. Bullet: sending feedback on contact points on depth 0 as well
     * [Pull request #2630](https://bitbucket.org/osrf/gazebo/pull-requests/2630/)
 
+1. Deprecate functions to set linear/angular acceleration 
+    * [Pull request #2622](https://bitbucket.org/osrf/gazebo/pull-request/2622)
+
 1. Added possibility to enforce contact computation
     * [Pull request #2629](https://bitbucket.org/osrf/gazebo/pull-requests/2629/)